Clean up and fix pagination and sorting
Pagination had some variables out in the open, moved them to
useEffect and useState. Fixed an issue where pressing next while on
the last page, it would still work and move out of the max pages.
Sorting was only working for the first press since the main array
was being sorted and not a placeholder array.